Декомпозирует компонент MainScreenContent
